refactor: Station 패키지 구조 상세화 및 클래스 이동
