+++ src/main/java/subway/line/entity/Line.java
@@ -1,6 +1,6 @@
 package subway.line.entity;
 
-import subway.station.Station;
+import subway.station.entity.Station;
 
 import javax.persistence.*;
