+++ src/main/java/subway/station/service/StationService.java
@@ -1,7 +1,11 @@
-package subway.station;
+package subway.station.service;
 
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import subway.station.dto.StationRequest;
+import subway.station.dto.StationResponse;
+import subway.station.entity.Station;
+import subway.station.repository.StationRepository;
 
 import java.util.List;
 import java.util.stream.Collectors;
